- moved component `SessionSummary` to be used as a UI and not have access to store 	- this will allow us to re-use the same component in multiple places
